% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_dog.R
\name{geom_dog}
\alias{geom_dog}
\title{geom_dog}
\usage{
geom_dog(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  inherit.aes = TRUE,
  na.rm = FALSE,
  by = "width",
  nudge_x = 0,
  ...
)
}
\arguments{
\item{mapping}{aes mapping}

\item{data}{data}

\item{stat}{stat}

\item{position}{position}

\item{inherit.aes}{logical, whether inherit aes from ggplot()}

\item{na.rm}{logical, whether remove NA values}

\item{by}{one of 'width' or 'height'}

\item{nudge_x}{horizontal adjustment to nudge dogs}

\item{...}{additional parameters}
}
\value{
geom layer
}
\description{
geom layer adding dogs
}
\examples{
library("ggplot2")
ggplot(mtcars) +
geom_dog(aes(mpg, wt), dog = "doge", size = 5)

set.seed(1)
df <- data.frame(x = rnorm(10),
                 y = rnorm(10),
                 image = sample(c("doge",
                                  "thisisfine",
                                  "tail",
                                  "chilaquil",
                                  "gabe"),
                                 size = 10, replace = TRUE))
 ggplot(df) +
geom_dog(aes(x, y, dog = image), size = 5)

}
