% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_dog.R
\name{geom_penguin}
\alias{geom_penguin}
\title{geom_penguin}
\usage{
geom_penguin(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  inherit.aes = TRUE,
  na.rm = FALSE,
  by = "width",
  nudge_x = 0,
  ...
)
}
\arguments{
\item{mapping}{aes mapping}

\item{data}{data}

\item{stat}{stat}

\item{position}{position}

\item{inherit.aes}{logical, whether inherit aes from ggplot()}

\item{na.rm}{logical, whether remove NA values}

\item{by}{one of 'width' or 'height'}

\item{nudge_x}{horizontal adjustment to nudge penguins}

\item{...}{additional parameters}
}
\value{
geom layer
}
\description{
geom layer adding penguins
}
\examples{
library("ggplot2")
ggplot(mtcars) +
geom_penguin(aes(mpg, wt), penguin = "Adelie-allison_horst", size = 5)

set.seed(1)
df <- data.frame(x = rnorm(10),
                 y = rnorm(10),
                 image = sample(c("Adelie-allison_horst",
                                  "chinstrap-allison_horst",
                                  "gentoo-allison_horst"),
                                 size = 10, replace = TRUE))
 ggplot(df) +
geom_penguin(aes(x, y, penguin = image), size = 5)

}
